+++ src/game/shared/achievementmgr.cpp
@@ -141,8 +141,8 @@ static void WriteAchievementGlobalState(KeyValues *pKV, bool bPersistToSteamClou
                                               : NULL;
 
         if (pRemoteStorage) {
-            uint64 availableBytes = 0;
-            uint64 totalBytes = 0;
+            int32 availableBytes = 0;
+            int32 totalBytes = 0;
             if (pRemoteStorage->GetQuota(&totalBytes, &availableBytes)) {
                 if (totalBytes > 0) {
                     int32 filesize = (int32) filesystem->Size(szFilename);
